Revert to dynamic linking on Windows, bundle DLLs with artifact
- Use x64-windows triplet (dynamic) instead of x64-windows-static
- Copy vcpkg runtime DLLs to Release folder before uploading artifact
- Add MSVC warning suppression for generated protobuf code (C4100, C4127)
- Fix size_t to uint32_t conversion warning in VideoService::GetConfig

+++ src/service/src/VideoService.cpp
@@ -103,8 +103,8 @@ grpc::Status VideoServiceImpl::GetConfig(
     const GetConfigRequest* /*request*/,
     VideoConfig* response) {
 
-    response->set_width(frame_buffer_.width());
-    response->set_height(frame_buffer_.height());
+    response->set_width(static_cast<uint32_t>(frame_buffer_.width()));
+    response->set_height(static_cast<uint32_t>(frame_buffer_.height()));
     response->set_framerate_hz(50);  // PAL
 
     return grpc::Status::OK;
